Multiple changes:
1. Activated optimizations in 'apply-an.sh' script. Optimizations
   at (roughly) "-O2" are now applied by both clang and the encode
   executable.

2. Proper handling of calls to external functions: Function
   arguments are decoded before being passes, and return values
   are encoded.

3. Added encoding of further binary operations.
   (cf. 'OperationsEncoder.cpp')

4. Handling of 'ptrtoint' and 'inttoptr' casts: Currently pointers
   are not encoded. Therefore encoding/decoding operations must be
   inserted when we cast between pointers and integers. (At the
   moment this handling is done by the 'GEPHandler', but it would
   probably be logically more consistent if the 'OperationsEncoder'
   did this.)

5. Added a test case for the handling of 'ptrtoint' and 'inttoptr'
   casts (i.e. 'ptrarith.c').

+++ encode/runtime/rdtsc.c
@@ -12,7 +12,7 @@ uint64_t rdtsc() {
                         ::: "%eax", "%ebx", "%ecx", "%edx" );
 #endif
   __asm__ __volatile__ ("rdtsc" : "=a" (low), "=d" (high));
-  return (uint64_t)high << 32 | low;
+  return (((uint64_t)high) << 32) | (uint64_t)low;
 }
 
 void cycles_msg(uint64_t cs) {
